refactor: landing 파일 분리
- 랜딩 페이지 관련 코드가 너무 길어서
- 섹션별로 나눠서 별도 폴더에 저장
- 스타일도 별도 파일로 분리
